move agent & serial agent to kernel; deprecate Agents & SerialAgents in favour of (Serial)Agent.Make to match cats-effect style; add carriers to interop with ce2; drop MakeRef and MakeSemaphore for implicit instances; add package object concurrent into kernel;
