+++ modules/kernel/src/main/scala/tofu/internal/carriers/concurrent.scala
@@ -1,8 +1,7 @@
 package tofu.internal.carriers
 
 import tofu.internal.Interop
-import tofu.concurrent.MakeAtom
-import tofu.concurrent.MakeQVar
+import tofu.concurrent.{MakeAgent, MakeAtom, MakeQVar, MakeSerialAgent}
 
 trait MkAtomCE2Carrier[I[_], F[_]] extends MakeAtom[I, F]
 
@@ -30,4 +29,30 @@ trait MkQVarCE3Carrier[I[_], F[_]] extends MakeQVar[I, F]
 object MkQVarCE3Carrier {
   final implicit def interopCE3Carrier[I[_], F[_]]: MkQVarCE3Carrier[I, F] =
     macro Interop.delegate2[MkQVarCE3Carrier[I, F], I, F, { val `tofu.interop.CE3Kernel.qvarByConcurrent`: Unit }]
+}
+
+trait MkAgentCE2Carrier[I[_], F[_]] extends MakeAgent[I, F]
+
+object MkAgentCE2Carrier {
+  final implicit def interopCE2Carrier[I[_], F[_]]: MkAgentCE2Carrier[I, F] =
+    macro Interop
+      .delegate2[MkAgentCE2Carrier[I, F], I, F, { val `tofu.interop.CE2Kernel.agentByRefAndSemaphore`: Unit }]
+}
+
+trait MkSerialAgentCE2Carrier[I[_], F[_]] extends MakeSerialAgent[I, F]
+
+object MkSerialAgentCE2Carrier extends MkSerialAgentCE2Carrier0 {
+  final implicit def interopCE2Carrier[I[_], F[_]]: MkSerialAgentCE2Carrier[I, F] =
+    macro Interop.delegate2[MkSerialAgentCE2Carrier[
+      I,
+      F
+    ], I, F, { val `tofu.interop.CE2Kernel.serialAgentByRefAndSemaphore`: Unit }]
+}
+trait MkSerialAgentCE2Carrier0 {
+  final implicit def underlyinginteropCE2Carrier[I[_], F[_]]: MkSerialAgentCE2Carrier[I, F] =
+    macro Interop.delegate2[MkSerialAgentCE2Carrier[
+      I,
+      F
+    ], I, F, { val `tofu.interop.CE2Kernel.underlyingSerialAgentByRefAndSemaphore`: Unit }]
+
 }
