+++ modules/kernelCE2Interop/src/main/scala/tofu/interop/CE2Kernel.scala
@@ -1,13 +1,14 @@
 package tofu
 package interop
 
-import cats.effect.concurrent.{MVar, Ref}
+import cats.effect.concurrent.{MVar, Ref, Semaphore}
 import cats.effect.{Async, Blocker, Bracket, Concurrent, ContextShift, Effect, ExitCase, Fiber, IO, Sync, Timer}
-import cats.{Id, ~>}
+import cats.{FlatMap, Id, Monad, effect, ~>}
 import tofu.compat.unused
-import tofu.concurrent.{Atom, QVar}
+import tofu.concurrent.{Agent, Atom, MakeAgent, MakeSerialAgent, QVar, SerialAgent}
 import tofu.internal.NonTofu
 import tofu.internal.carriers._
+import tofu.lift.Lift
 import tofu.syntax.monadic._
 
 import scala.concurrent.duration.FiniteDuration
@@ -107,5 +108,39 @@ object CE2Kernel {
   final def sleep[F[_]](implicit T: cats.effect.Timer[F]): SleepCE2Carrier[F] =
     new SleepCE2Carrier[F] {
       def sleep(duration: FiniteDuration): F[Unit] = T.sleep(duration)
+    }
+
+  final def agentByRefAndSemaphore[I[_]: Sync, F[_]](implicit F: Concurrent[F]): MakeAgent[I, F] = {
+    implicit val fire: Fire[F] = new Fire[F] {
+      override def fireAndForget[A](fa: F[A]): F[Unit] = F.void(F.start(fa))
+    }
+    new MakeAgent[I, F] {
+      def agentOf[A](a: A): I[Agent[F, A]] =
+        for {
+          ref <- Ref.in[I, F, A](a)
+          sem <- Semaphore.in[I, F](1)
+        } yield SemRef(ref, sem)
+    }
+  }
+
+  final def serialAgentByRefAndSemaphore[I[_]: Sync, F[_]](implicit F: Concurrent[F]): MakeSerialAgent[I, F] =
+    new MakeSerialAgent[I, F] {
+      override def serialAgentOf[A](a: A): I[SerialAgent[F, A]] =
+        for {
+          ref <- Ref.in[I, F, A](a)
+          sem <- Semaphore.in[I, F](1)
+        } yield SerialSemRef(ref, sem)
+    }
+
+  final def underlyingSerialAgentByRefAndSemaphore[I[_]: Sync, F[_], G[_]: Monad](implicit
+      F: Concurrent[F],
+      lift: Lift[F, G]
+  ): MakeSerialAgent[I, G] =
+    new MakeSerialAgent[I, G] {
+      override def serialAgentOf[A](a: A): I[SerialAgent[G, A]] =
+        for {
+          ref <- Ref.in[I, F, A](a)
+          sem <- Semaphore.in[I, F](1)
+        } yield UnderlyingSemRef[F, G, A](ref, sem)
     }
 }
